+++ src/panel_material_ui/widgets/base.py
@@ -4,7 +4,6 @@ from typing import TYPE_CHECKING, TypeVar
 
 import param
 from panel._param import Margin
-from panel.viewable import Viewable
 from panel.widgets.base import WidgetBase
 
 from ..base import ESMTransform, MaterialComponent
@@ -96,29 +95,3 @@ class MaterialWidget(MaterialComponent, WidgetBase):
         if isinstance(parameter.owner, MaterialComponent):
             widget.jslink(parameter.owner, value=parameter.name)
         return widget
-
-    def api(self, jslink: bool=False, sizing_mode="stretch_width", **kwargs)->Viewable:
-        """Returns an interactive component for exploring the API of the widget.
-
-        Parameters
-        ----------
-        jslink: bool
-            Whether to use jslinks instead of Python based links.
-            This does not allow using all types of parameters.
-        sizing_mode: str
-            Sizing mode for the component.
-        kwargs: dict
-            Additional arguments to pass to the component.
-
-        Example:
-        --------
-        >>> pmui.Button(name="Open").api()
-        """
-        import panel as pn
-
-        import panel_material_ui as pmui
-        return pmui.Tabs(
-            pn.pane.HTML(self.param, name="Parameter Table", sizing_mode="stretch_width"),
-            pmui.Row(self.controls(jslink=jslink), self, name="Parameter Editor", sizing_mode="stretch_width"),
-            sizing_mode=sizing_mode, **kwargs
-        )
